Initialize self.logdir also when 'logdir' is empty
In the costructor of the BaseExperiment class, self.logdir can be None
if the env_config file does not contain a 'logdir' key. This is a
problem because the library user cannot get the log directory. Solution:
use pathlib to get a default logdir from home directory, then
'ray_results'.