Allow BaseExperiment to accept experiment configuration directly (#13)
* Allow BaseExperiment to accept exp_config directly

This commit adds the ability to pass the exp_config dictionary to
BaseExperiment instead of a path to a JSON file on disk. Note that there
will be `complete_config/exp_config.json` in the results anyway.

* Ignore Vim swap files

* Allow exp_config to have env and ray config

This commit adds the ability to pass ray and env configurations
directly, rather than using files. In any case, when the experiment is
run, the three files (ray, exp and env configs) will be in the
`complete_config` directory.

* Document '(ray|exp|env)_config'

* Allow TuningExperiment to accept tune_config

* Document '(ray|exp|env)_config' in README

* Fix documentation (warnings and typos)

* Rewrite not_defined() as "not defined(...)"